Design: Update salary input
