--- src/api/axios.ts
@@ -1,13 +1,0 @@
-import { removeCookieToken } from '@/libs/Cookie';
-import axios from 'axios';
-// import { onRequest, onRequestError } from '../libs/refresh';
-
-export const instance = axios.create({
-  baseURL: 'https://kingtaeyoon.shop/api',
-  // timeout: 2500,
-  headers: {
-    'Content-type': 'application/json',
-  },
-});
-
-// instance.interceptors.request.use(onRequest, onRequestError);
